data-generator/requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
Faker==22.0.0
python-dotenv==1.0.0
